requirements_ingestion.txt
openai==1.3.0
pymongo==4.6.0
tiktoken==0.5.1
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
streamlit==1.28.1
plotly==5.17.0
